+++ api/service/error_test.go
@@ -1,0 +1,12 @@
+package service
+
+import (
+	"testing"
+
+	"github.com/stvp/assert"
+)
+
+func TestMissingExecutionError(t *testing.T) {
+	e := MissingExecutionError{ID: "test"}
+	assert.Equal(t, "Execution test doesn't exists", e.Error())
+}