+++ cmd/service/utils.go
@@ -3,7 +3,10 @@ package service
 import (
 	"fmt"
 	"io/ioutil"
+	"net/url"
 	"os"
+
+	"gopkg.in/src-d/go-git.v4/plumbing"
 
 	"github.com/asaskevich/govalidator"
 	"github.com/logrusorgru/aurora"
@@ -12,12 +15,17 @@ import (
 	"github.com/mesg-foundation/core/config"
 	"github.com/mesg-foundation/core/container"
 	"github.com/mesg-foundation/core/service"
+	"github.com/mesg-foundation/core/service/importer"
 	"github.com/spf13/viper"
 	"google.golang.org/grpc"
 	git "gopkg.in/src-d/go-git.v4"
 )
 
-var cli core.CoreClient
+func cli() core.CoreClient {
+	connection, err := grpc.Dial(viper.GetString(config.APIClientTarget), grpc.WithInsecure())
+	utils.HandleError(err)
+	return core.NewCoreClient(connection)
+}
 
 // Set the default path if needed
 func defaultPath(args []string) string {
@@ -27,30 +35,33 @@ func defaultPath(args []string) string {
 	return "./"
 }
 
+func handleValidationError(err error) {
+	if _, ok := err.(*importer.ValidationError); ok {
+		fmt.Println(aurora.Red(err))
+		fmt.Println("Run the command 'service validate' for more details")
+		os.Exit(0)
+	}
+}
+
 // prepareService downloads if needed, create the service, build it and inject configuration
 func prepareService(path string) (importedService *service.Service) {
 	path, didDownload, err := downloadServiceIfNeeded(path)
 	utils.HandleError(err)
 	if didDownload {
 		defer os.RemoveAll(path)
-		fmt.Println(aurora.Green("Service downloaded with success"))
-		fmt.Println("Temp folder: " + path)
+		fmt.Printf("%s Service downloaded with success\n", aurora.Green("✔"))
 	}
-	importedService, err = service.ImportFromPath(path)
-	if err != nil {
-		fmt.Println("Run the command 'service validate' to get detailed errors")
-		utils.HandleError(err)
-	}
+	importedService, err = importer.From(path)
+	handleValidationError(err)
+	utils.HandleError(err)
 	imageHash, err := buildDockerImage(path)
 	utils.HandleError(err)
-	fmt.Println(aurora.Green("Image built with success"))
-	fmt.Println("Image hash:", imageHash)
+	fmt.Printf("%s Image built with success\n", aurora.Green("✔"))
 	injectConfigurationInDependencies(importedService, imageHash)
 	return
 }
 
 func downloadServiceIfNeeded(path string) (newPath string, didDownload bool, err error) {
-	didDownload = false
 	newPath = path
 	if govalidator.IsURL(path) {
 		newPath, err = createTempFolder()
@@ -58,16 +69,27 @@ func downloadServiceIfNeeded(path string) (newPath string, didDownload bool, err
 			return
 		}
 		err = gitClone(path, newPath, "Downloading service...")
-		didDownload = true
+		didDownload = err == nil
 	}
 	return
 }
 
-func gitClone(url string, path string, message string) (err error) {
+func gitClone(repoURL string, path string, message string) (err error) {
+	u, err := url.Parse(repoURL)
+	if err != nil {
+		return err
+	}
+	if u.Scheme == "" {
+		u.Scheme = "https"
+	}
+	options := &git.CloneOptions{}
+	if u.Fragment != "" {
+		options.ReferenceName = plumbing.ReferenceName("refs/heads/" + u.Fragment)
+		u.Fragment = ""
+	}
+	options.URL = u.String()
 	utils.ShowSpinnerForFunc(utils.SpinnerOptions{Text: message}, func() {
-		_, err = git.PlainClone(path, false, &git.CloneOptions{
-			URL: url,
-		})
+		_, err = git.PlainClone(path, false, options)
 	})
 	return
 }
@@ -100,10 +122,4 @@ func injectConfigurationInDependencies(s *service.Service, imageHash string) {
 		s.Dependencies = make(map[string]*service.Dependency)
 	}
 	s.Dependencies["service"] = dependency
-}
-
-func init() {
-	connection, err := grpc.Dial(viper.GetString(config.APIClientTarget), grpc.WithInsecure())
-	utils.HandleError(err)
-	cli = core.NewCoreClient(connection)
 }