+++ scripts/update-doc.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+echo "Generate API documentation"
+./scripts/build-proto.sh
+echo "Generate CLI documentation"
+go run docs/generate.go